Move HierarchicalLinearModelDefinition together with StaticHierarchicalLinearModel and tidy.
